refactor: remove explore.tsx and related logic
Remove the explore.tsx file and all associated radar data 
fetching and clustering logic to streamline the application. 
This change simplifies the codebase and prepares for future 
refactoring efforts.
