+++ app/(tabs)/explore/_layout.tsx
@@ -1,0 +1,16 @@
+import { Stack } from "expo-router";
+
+export default function ExploreLayout() {
+    return (
+        <Stack>
+            <Stack.Screen name="map" options={{ title: "Map", headerShown: false }} />
+            <Stack.Screen 
+                name="search" 
+                options={{ 
+                    title: "Search",
+                    presentation:"modal"
+                }} 
+            />
+        </Stack>
+    )
+}
